Moved 'Mine' and 'Block' (#186)
