+++ src/me/sirfaizdat/prison/mines/Mine.java
@@ -17,10 +17,11 @@
  *
  */
 
-package me.sirfaizdat.prison.mines.entities;
+package me.sirfaizdat.prison.mines;
 
 import me.sirfaizdat.prison.core.Prison;
 import me.sirfaizdat.prison.json.JsonMine;
+import me.sirfaizdat.prison.mines.Block;
 import me.sirfaizdat.prison.mines.SerializableMine;
 import me.sirfaizdat.prison.mines.events.MineResetEvent;
 import org.bukkit.Bukkit;
