+++ src/me/sirfaizdat/prison/mines/MinesManager.java
@@ -20,8 +20,6 @@ package me.sirfaizdat.prison.mines;
 
 import me.sirfaizdat.prison.core.Prison;
 import me.sirfaizdat.prison.json.JsonMine;
-import me.sirfaizdat.prison.mines.entities.Block;
-import me.sirfaizdat.prison.mines.entities.Mine;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.World;
