+++ src/me/sirfaizdat/prison/mines/cmds/CmdRemoveBlock.java
@@ -23,8 +23,8 @@ import me.sirfaizdat.prison.core.ItemManager.ItemSet;
 import me.sirfaizdat.prison.core.MessageUtil;
 import me.sirfaizdat.prison.core.Prison;
 import me.sirfaizdat.prison.mines.Mines;
-import me.sirfaizdat.prison.mines.entities.Block;
-import me.sirfaizdat.prison.mines.entities.Mine;
+import me.sirfaizdat.prison.mines.Block;
+import me.sirfaizdat.prison.mines.Mine;
 import org.bukkit.Material;
 
 /**
